feat(cli): add check for Gateway API in `linkerd check --pre`
This checks if the GW API is already installed. If not, it just issues a warning and doesn't fail, to account for users wishing to manage it via Linkerd itself.
It realies on the same function introduced in #13834 which has been moved to the new `pkg/healthcheck/gwapi` package for reusability.

```bash
$ bin/linkerd check --pre
kubernetes-api
--------------
√ can initialize the client
√ can query the Kubernetes API

kubernetes-version
------------------
√ is running the minimum Kubernetes API version

pre-kubernetes-setup
--------------------
√ control plane namespace does not already exist
‼ Gateway API is installed
    The Gateway API CRDs should be pre-installed, unless Linkerd is explicitly set to deploy them
    see https://linkerd.io/2/checks/#gwapi for hints
√ can create non-namespaced resources
√ can create ServiceAccounts
√ can create Services
√ can create Deployments
√ can create CronJobs
√ can create ConfigMaps
√ can create Secrets
√ can read Secrets
√ can read extension-apiserver-authentication configmap
√ no clock skew detected

linkerd-version
---------------
√ can determine the latest version
‼ cli is up-to-date
    unsupported version channel: dev-a1d7ff19-alpeb
    see https://linkerd.io/2/checks/#l5d-version-cli for hints

Status check results are √
```
