+++ cli/cmd/install.go
@@ -20,6 +20,7 @@ import (
 	pkgcmd "github.com/linkerd/linkerd2/pkg/cmd"
 	flagspkg "github.com/linkerd/linkerd2/pkg/flags"
 	"github.com/linkerd/linkerd2/pkg/healthcheck"
+	"github.com/linkerd/linkerd2/pkg/healthcheck/gwapi"
 	"github.com/linkerd/linkerd2/pkg/k8s"
 	"github.com/linkerd/linkerd2/pkg/tree"
 	"github.com/spf13/cobra"
@@ -205,39 +206,6 @@ func checkNoConfig(ctx context.Context, k8sAPI *k8s.KubernetesAPI) error {
 	return nil
 }
 
-type GatewayAPICRDs int
-
-const (
-	Absent GatewayAPICRDs = iota
-	Linkerd
-	External
-)
-
-// checkGatewayAPICRDs returns true if the Gateway API CRDs are installed in the
-// cluster, and false otherwise.
-func checkGatewayAPICRDs(ctx context.Context, k8sAPI *k8s.KubernetesAPI) (GatewayAPICRDs, error) {
-	crds := k8sAPI.Apiextensions.ApiextensionsV1().CustomResourceDefinitions()
-	result := Absent
-	names := []string{
-		"httproutes.gateway.networking.k8s.io",
-		"grpcroutes.gateway.networking.k8s.io",
-	}
-	for _, name := range names {
-		crd, err := crds.Get(ctx, name, metav1.GetOptions{})
-		if err == nil && crd != nil {
-			if crd.Annotations[k8s.CreatedByAnnotation] != "" {
-				return Linkerd, nil
-			}
-			result = External
-		} else if kerrors.IsNotFound(err) {
-			// No action if CRD is not found.
-		} else {
-			return Absent, err
-		}
-	}
-	return result, nil
-}
-
 func installCRDs(ctx context.Context, k8sAPI *k8s.KubernetesAPI, w io.Writer, options valuespkg.Options, format string) error {
 	if err := checkNoConfig(ctx, k8sAPI); err != nil {
 		return err
@@ -365,14 +333,14 @@ func renderChartToBuffer(files []*loader.BufferedFile, values map[string]interfa
 	return &buf, vals, nil
 }
 
-func updateDefaultValues(installed GatewayAPICRDs, defaultValues map[string]interface{}) map[string]interface{} {
-	if installed == Absent {
+func updateDefaultValues(installed gwapi.GatewayAPICRDs, defaultValues map[string]interface{}) map[string]interface{} {
+	if installed == gwapi.Absent {
 		// if GW API is not installed, default to false
 		defaultValues["installGatewayAPI"] = false
-	} else if installed == Linkerd {
+	} else if installed == gwapi.Linkerd {
 		// if it is installed by Linkerd, default to true
 		defaultValues["installGatewayAPI"] = true
-	} else if installed == External {
+	} else if installed == gwapi.External {
 		// if it is external, default to false as we are not managing it
 		defaultValues["installGatewayAPI"] = false
 	}
@@ -380,7 +348,7 @@ func updateDefaultValues(installed GatewayAPICRDs, defaultValues map[string]inte
 	return defaultValues
 }
 
-func validateFinalValues(installed GatewayAPICRDs, finalValues map[string]interface{}) error {
+func validateFinalValues(installed gwapi.GatewayAPICRDs, finalValues map[string]interface{}) error {
 	installing := false
 
 	if installGatewayAPI, ok := finalValues["installGatewayAPI"]; ok {
@@ -391,17 +359,17 @@ func validateFinalValues(installed GatewayAPICRDs, finalValues map[string]interf
 		installing = enableHttpRoutes == true
 	}
 
-	if installed == Absent {
+	if installed == gwapi.Absent {
 		if !installing {
 			// if we are not installing GW API Resources and they are not present, error
 			return errors.New("The Gateway API CRDs must be installed prior to installing Linkerd: https://gateway-api.sigs.k8s.io/guides/#installing-gateway-api")
 		}
-	} else if installed == Linkerd {
+	} else if installed == gwapi.Linkerd {
 		if !installing {
 			// if they are installed and managed by Linkerd, we cannot uninstall them
 			return errors.New("Linkerd is providing GW API, but your current install configuration will remove it")
 		}
-	} else if installed == External {
+	} else if installed == gwapi.External {
 		if installing {
 			// if they are installed but are external, we cannot be installing as well
 			return errors.New("Linkerd cannot install the Gateway API CRDs because they are already installed by an external source. Please set `installGatewayAPI` to `false`.")
@@ -447,7 +415,7 @@ func renderCRDs(ctx context.Context, k *k8s.KubernetesAPI, w io.Writer, options 
 	// If any of the Gateway API CRDs are installed, we default to rendering the
 	// Gateway API CRDs.
 	if k != nil {
-		installed, err := checkGatewayAPICRDs(ctx, k)
+		installed, err := gwapi.CheckGatewayAPICRDs(ctx, k)
 		if err != nil {
 			return err
 		}
